+++ src/TemplateWhispererSuggestionListBuilder.php
@@ -4,9 +4,7 @@ namespace Drupal\template_whisperer;
 
 use Drupal\Core\Config\Entity\ConfigEntityListBuilder;
 use Drupal\Core\Entity\EntityInterface;
-use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeInterface;
-use Drupal\Core\Routing\UrlGeneratorInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
@@ -31,33 +29,14 @@ final class TemplateWhispererSuggestionListBuilder extends ConfigEntityListBuild
   protected $twSuggestionUsage;
 
   /**
-   * Constructs a TemplateWhispererSuggestionListBuilder object.
-   *
-   * @param \Drupal\Core\Entity\EntityTypeInterface $entity_type
-   *   The entity type definition.
-   * @param \Drupal\Core\Entity\EntityStorageInterface $storage
-   *   The entity storage class.
-   * @param \Drupal\Core\Routing\UrlGeneratorInterface $url_generator
-   *   The url generator service.
-   * @param \Drupal\template_whisperer\TemplateWhispererSuggestionUsage $tw_suggestion_usage
-   *   Template Whisperer Suggestion Usage.
-   */
-  public function __construct(EntityTypeInterface $entity_type, EntityStorageInterface $storage, UrlGeneratorInterface $url_generator, TemplateWhispererSuggestionUsage $tw_suggestion_usage) {
-    parent::__construct($entity_type, $storage);
-    $this->urlGenerator      = $url_generator;
-    $this->twSuggestionUsage = $tw_suggestion_usage;
-  }
-
-  /**
    * {@inheritdoc}
    */
   public static function createInstance(ContainerInterface $container, EntityTypeInterface $entity_type) {
-    return new self(
-      $entity_type,
-      $container->get('entity_type.manager')->getStorage($entity_type->id()),
-      $container->get('url_generator'),
-      $container->get('template_whisperer.suggestion.usage')
-    );
+    $instance                    = parent::createInstance($container, $entity_type);
+    $instance->urlGenerator      = $container->get('url_generator');
+    $instance->twSuggestionUsage = $container->get('template_whisperer.suggestion.usage');
+
+    return $instance;
   }
 
   /**